Render README.md and hyperlink it from the footer
Greetings!

I found it slightly nontrivial to figure out where the README lives at,
so I tweaked Jekyll to render it.  I hyperlinked the "README.md" text
found in each page's footer to this rendered copy.  I'm not a fan of
adding a Jekyll plugin to simply render a Markdown file without front
matter, but it does the trick.

Another alternative to using the jekyll-optional-front-matter plugin
is to simply add front-matter to the top of the README.md.  We could
try this, but I imagine at least one of Sourcehut, GitLab, or GitHub
will render the front-matter as a superfluous paragraph.  Additionally
Markdown is simply plain text, and can be read as such, so I don't
think it would be best to add line noise to the beginning of the
README.md.

Happy bike shedding!
winny

Signed-off-by: Tom Ryder <tom@sanctum.geek.nz>